requirements.txt
# Appium相关
Appium-Python-Client>=3.1.0
selenium>=4.15.0

# 测试框架
pytest>=7.4.0
pytest-html>=4.1.0
pytest-xdist>=3.3.0
pytest-rerunfailures>=12.0

# 日志和配置
loguru>=0.7.0
pyyaml>=6.0.1
python-dotenv>=1.0.0

# 工具库
requests>=2.31.0
allure-pytest>=2.13.0
pytest-reportlog>=0.3.0

# 数据处理
pandas>=2.0.0
openpyxl>=3.1.0

# 图像处理（用于截图对比）
Pillow>=10.0.0
opencv-python>=4.8.0

# 时间处理
python-dateutil>=2.8.2

# 类型提示
typing-extensions>=4.8.0
